update syntax for `new-window` and `set-option`
The `-a` flag for new window (to bump window indices) is not happy with
Tmux 3.0 when specifying a not-yet-to-exist index. Removes the `-a`
flag, since we're calculating new window indices to spawn into anyway
and don't need to bump window indices.

Also updates `set-window-option` to use now-recommend `set-option -w`
syntax